+++ Messages/OMessages/companyfrozen.h
@@ -10,7 +10,7 @@ class CompanyFrozenResp : public OMessage
 {
     qint32 m_companyId;
 
-    qint32 length() const;
+    qint16 length() const;
 public:
     CompanyFrozenResp(qint32 companyId);
